Update sonarqube installation
I encountered an issue with installing the unzip package due to lack of root privileges. To resolve this, I switched to the sudo user and successfully installed the unzip package. Afterwards, I created a new user 'sonarqube' and then switched back to the sonarqube user to continue with the installation process.